PLEASDE

+++ assignment6-1.cpp
@@ -31,11 +31,11 @@ void prime(int begin, int end) // calculation function, returns 0 or 1
     int x, y;
     for (x = begin; x <= end; x++)
     {
-        for (y = 2; y <= (x / 2); y++){
+        for (y = 2; y <= (static_cast<float>(x) / 2); y++){
             if (x % y == 0)
                 break;
             }
-        if (y > (static_cast<double>(x)/2))
+        if (y > (static_cast<float>(x)/2))
             cout << "Prime numbers: " << x << endl;
             
     }
